xtasks: make `files_with_extension` more accessible
This function is useful beyond the formatting module, so put it into the
top level of the library.
